+++ src/DbPatch/Command/Remove.php
@@ -88,24 +88,24 @@ class DbPatch_Command_Remove extends DbPatch_Command_Abstract
 
     /**
      * Remove patch from the changelog table
-     * 
+     *
      * @param int $patchNumber
      * @param string $branchName
      * @return void
      */
     protected function removePatch($patchNumber, $branchName)
     {
-        $db = $this->getDb();
+        $db = $this->getDb()->getAdapter();
         $branchSQL = "";
 
         if (!empty($branchName)) {
-            $branchSQL = sprintf("AND `branch` = '%s'",
+            $branchSQL = sprintf("AND branch = '%s'",
                                  $branchName);
         }
 
         $query = sprintf("SELECT branch
-                              FROM `%s`
-                              WHERE `patch_number` = %d {$branchSQL}",
+                              FROM %s
+                              WHERE patch_number = %d {$branchSQL}",
                          self::TABLE,
                          $patchNumber);
 
@@ -130,13 +130,12 @@ class DbPatch_Command_Remove extends DbPatch_Command_Abstract
         else {
             $branchMsg = (empty($branchName) ? ""
                     : "from branch '$branchName' ");
-            $query = sprintf("DELETE FROM `%s`
-                                  WHERE `patch_number` = %d {$branchSQL}",
+            $query = sprintf("DELETE FROM %s
+                                  WHERE patch_number = %d {$branchSQL}",
                              self::TABLE,
                              $patchNumber);
 
             $db->query($query);
-            $db->commit();
             $this->getWriter()->line("Removed patch $patchNumber {$branchMsg}in the `" . self::TABLE . "` table");
         }
     }