+++ CHANGES.md
@@ -1,10 +1,41 @@
-+ : new feature
-- : deprecated feature
-* : bugfix or changed code
-Bug numbers are references to bugs in https://github.com/dbpatch/DbPatch/issues
+\+ : new feature
+
+\- : deprecated feature
+
+\* : bugfix or changed code
+
+Issue numbers are references to issues in [https://github.com/dbpatch/DbPatch/issues](https://github.com/dbpatch/DbPatch/issues)
+
 Please do not exceed 80 characters per line.
 Latest changes go on top, different types of changes may be mixed.
 
-Changes in DbPatch 1.0
-----------------------
-+  Initial version
+
+Changes in DbPatch 1.1.0
+------------------------
+\+ Added support for PHP Object patches
+
+\+ Added correct exit status
+
+\* Fixed bug in the info command
+
+\+ Added support for Sqlite3 and PostgreSQL
+
+\+ Added setup command
+
+\+ Added composer support
+
+\* Optimized status and show commands (less memory usage)
+
+Changes in DbPatch 1.0.1
+------------------------
+\+ Added info command
+
+\+ Added dump directory config
+
+\* Fixed detection of trailing slash in paths
+
+Changes in DbPatch 1.0.0
+------------------------
+\+  Initial version
+
+